move calculate_cluster_factors to cluster assign to allow splitting it out
